dvc: refactor cache/remote relations
